% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botornot.R
\name{botornot}
\alias{botornot}
\title{botornot}
\usage{
botornot(x)
}
\arguments{
\item{x}{Object to be classified. Can be user IDs, screen names, or
data frames returned by rtweet.}
}
\value{
Classifications for all users expressed as probability of whether
  each account is a bot.
}
\description{
Classify users/accounts in Twitter data as bots or not bots.
}
