% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{botornot_model}
\alias{botornot_model}
\title{botornot model}
\format{A gbm model object}
\usage{
botornot_model
}
\description{
The model used to generate predicted probabilities.
}
\examples{
botornot_model
}
\keyword{datasets}
