% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{classify_data}
\alias{classify_data}
\title{classify data}
\usage{
classify_data(x, model)
}
\arguments{
\item{x}{New data on which to apply botornot model.}

\item{model}{gbm model from which to predict.}
}
\value{
Vector of predictions expressed as probabilities of accounts being
bots.
}
\description{
Generate predicted probabilities of observations being bots.
}
