% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet_join.R
\name{rtweet_join}
\alias{rtweet_join}
\title{retweet_join}
\usage{
rtweet_join(x)
}
\arguments{
\item{x}{Data returned by rtweet function to be merged (users and tweets)}
}
\value{
A single tibble with both users and tweets data
}
\description{
retweet_join
}
