% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetbotornot.R
\name{tweetbotornot}
\alias{tweetbotornot}
\alias{botornot}
\alias{botornot_dbl}
\title{tweetbotornot}
\usage{
tweetbotornot(x, fast = FALSE)

botornot(x, fast = FALSE)

botornot_dbl(x, fast = TRUE)
}
\arguments{
\item{x}{Object to be classified. Can be user IDs, screen names, or
data frames returned by rtweet.}

\item{fast}{Logical indicating whether to use the fast (lighter) model. The
default (fast = FALSE) method uses the most recent 100 tweets posted by
users to determine the probability of bot. The fast (fast = TRUE) method
only uses users-level data, which is easier to get in large quantities from
Twitter's APIS but overall less accurate.}
}
\value{
Classifications for all users expressed as probability of whether
each account is a bot.

A named (screen names or user IDs depending on input) numeric
vector of probabilities
}
\description{
Classify users/accounts in Twitter data as bots or not bots.

Returns a numeric vector of probabilities
}
\examples{
\dontrun{

## vector of screen names
sns <- c("kearneymw", "geoffjentry", "p_barbera",
  "tidyversetweets", "rstatsbot1234", "RStatsStExBot")

## get and view bot probability estimates
twb <- tweetbotornot(sns)
twb

## ask for the fast (user-level data only) version
twbf <- tweetbotornot(sns, fast = TRUE)
twbf

}
}
