% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botrnot.R
\name{botornot}
\alias{botornot}
\alias{botrnot}
\title{botornot}
\usage{
botornot(x, fast = FALSE)

botrnot(...)
}
\arguments{
\item{x}{Object to be classified. Can be user IDs, screen names, or
data frames returned by rtweet.}

\item{fast}{Logical indicating whether to use the fast (lighter) model. The
default (fast = FALSE) method uses the most recent 100 tweets posted by
users to determine the probability of bot. The fast (fast = TRUE) method
only uses users-level data, which is easier to get in large quantities from
Twitter's APIS but overall less accurate.}
}
\value{
Classifications for all users expressed as probability of whether
each account is a bot.
}
\description{
Classify users/accounts in Twitter data as bots or not bots.
}
